requirements.txt
streamlit>=1.18
google-generativeai>=0.5.0
requests>=2.28
python-dotenv>=1.0
flask>=2.0.0
flask-cors>=3.0.0